Created base template for integrating dotnetcore, webpack, typescript, and react
This template will be used for dotnetcore 2.1 MVC web apps, webpack 4, and the latest versions of typescript and react.

This integration will assist in starting up new projects and supporting existing projects.  I want to ensure that we can leverage webpack for best-in-class bundling, typescript for development assitance, and react for a minimal view framework.